Move shape checking utilities to private module
