+++ sympy/matrices/expressions/_shape.py
